Fix Vulkan pipeline setup

+++ 05b_CommandBuffersOverview/first_app.cpp
@@ -63,7 +63,7 @@ void FirstApp::createCommandBuffers() {
     throw std::runtime_error("failed to allocate command buffers");
   }
 
-  for (int i = 0; i < commandBuffers.size(); i++) {
+  for (size_t i = 0; i < commandBuffers.size(); i++) {
     VkCommandBufferBeginInfo beginInfo{};
     beginInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
